+++ CHANGES.rst
@@ -6,3 +6,7 @@ Changelog of lizard-history
 ----------------
 
 - Initial library skeleton created by nensskel.  [Arjan Verkerk]
+
+- Added django-tls to be able to use the request for logging user
+
+- Added receiver that logs changes to django models in django admin's logentry